// data/github/current/ui/packages/pull-request-viewer/components/__generated__/PullRequestMainContentAreaQuery.graphql.ts
export type PullRequestMainContentAreaQuery$variables = {
  endOid?: string | null | undefined;
  number: number;
  owner: string;
  repo: string;
  singleCommitOid?: string | null | undefined;
  startOid?: string | null | undefined;
};
export type PullRequestMainContentAreaQuery$data = {
  readonly repository: {
    readonly pullRequest: {
      readonly comparison: {
        readonly oldCommit: {
          readonly oid: any;
        };
      } | null | undefined;
      readonly " $fragmentSpreads": FragmentRefs<"Navigation_pullRequest">;
    } | null | undefined;
  } | null | undefined;
};
export type PullRequestMainContentAreaQuery = {
  response: PullRequestMainContentAreaQuery$data;
  variables: PullRequestMainContentAreaQuery$variables;
};